chore: split files to folders Signed-off-by: Florent Benoit <fbenoit@redhat.com>
