+++ packages/renderer/src/lib/task-manager/button/TaskManagerBulkDeleteButton.spec.ts
