+++ packages/renderer/src/lib/task-manager/button/TaskManagerClearAllButton.svelte
