+++ packages/renderer/src/lib/task-manager/table/TaskManagerTableProgressColumnCompleted.spec.ts
