+++ packages/renderer/src/lib/task-manager/table/TaskManagerTableProgressColumnCompleted.svelte
