+++ packages/renderer/src/lib/task-manager/table/action/TaskManagerActionsCancel.svelte
