+++ packages/renderer/src/lib/task-manager/table/action/TaskManagerActionsDelete.spec.ts
