+++ packages/renderer/src/lib/task-manager/table/action/TaskManagerActionsDelete.svelte
